refactor(cli): extract plugins command module; slim cli/bakin.ts to a thin router
The plugins module owns the whiskit publish/artifacts-index, plugin
import-export, and node:fs static imports — the CLI entry's static import
graph is now just http/output/help + APP_VERSION; every command group
loads lazily from its switch case.
